move _ForwardRedirection to the models/utils
